pyproject.toml
[project]
name = "devbot2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "chonkie>=1.0.10",
    "openai>=1.88.0",
    "pydantic>=2.11.5",
    "python-dotenv>=1.1.0",
    "robyn>=0.68.0",
    "celery>=5.3.6",
    "redis>=5.0.4",
]
